requirements.txt
fastapi
uvicorn
alembic
sqlalchemy
aiohttp
psycopg2-binary
python-dotenv
python-multipart
openpyxl
pandas
